Port over from Trial Search to Menu Header
